1、贷款平台管理，查和删完成，增和改有问题

+++ api/models/Platform.js
@@ -27,7 +27,7 @@ module.exports = {
     },
     grade: {
       type: 'number',
-      columnType:'int',
+      columnType:'float',
     },
     fastestTime: {
       type: 'string'
